[Fix] Handle different way to access array reference, [Change] Move protected helper helpers into trait,
